cleanup Dialect test
