+++ org.eclipse.daanse.db/dialect/db/netezza/pom.xml
@@ -26,5 +26,11 @@
       <groupId>biz.aQute.bnd</groupId>
       <artifactId>biz.aQute.bndlib</artifactId>
     </dependency>
+        <dependency>
+      <groupId>org.eclipse.daanse</groupId>
+      <artifactId>org.eclipse.daanse.db.dialect.db.postgresql</artifactId>
+      <version>${revision}</version>
+      <scope>compile</scope>
+    </dependency>
   </dependencies>
 </project>
