+++ org.eclipse.daanse.db/dialect/db/netezza/src/test/java/org/eclipse/daanse/db/dialect/db/netezza/ServiceTest.java
